make by ARSA group
create a simple website where payment gateway is integrated.
There will be a simple donate button on homepage. on clicking the donate button , the user will land on the payment type, eg. credit card , paypal, etc.
Once the payment is done and invoice will be generated and email will be sent to the user for the payment received. The invoice will contain he amount.
on any page / email, only basic information is needed. 
Create your own temporary / sandbox / testing accounts with 3rd party for integrations.
it is the task -3 f the spark foundation